refactor: extract useBattlePanel controller, make index a pure view
